+++ src/habutton.cpp
@@ -6,10 +6,10 @@
 const char *HAButton::component PROGMEM  = "button";
 
 HAButton::HAButton(const char *unique_id, const char *name, HADevice& device):
-    HAEntity(unique_id,name,&device,component) {
+    HAEntity(unique_id,name,component,&device) {
 }
 HAButton::HAButton(const char *unique_id, const char *name):
-    HAEntity(unique_id,name,NULL,component) {
+    HAEntity(unique_id,name,component) {
 }
 
 void HAButton::onConnect(PubSubClient * client){