+++ src/haentity.h
@@ -5,8 +5,11 @@
 #define __HAENTITY_H__
 
 #include <Arduino.h>
-class PubSubClient;
-class HADevice;
+#include "hakvpairlist.h"
+
+#define HA_FEATURE_AVAILABILITY -1
+#define HA_FEATURE_DEVICE_CLASS 0
+#define HA_FEATURE_ICON 1
 
 #ifndef HA_MAX_TOPIC_LENGTH
 #define HA_MAX_TOPIC_LENGTH 128
@@ -15,6 +18,8 @@ class HADevice;
 #define HA_MAX_PAYLOAD_LENGTH 512
 #endif
 
+class PubSubClient;
+class HADevice;
 
 /** Abstract class for HA entities
 
@@ -24,9 +29,9 @@ class HAEntity
     protected:
 
         static const char *configTopicTemplate;
-        static const char *configCommandTopicTemplate;
-        static const char *configStateTopicTemplate;
         static const char *configPayloadTemplate;
+        static const char *availabilityTopicTemplate;
+        static const char *featureKeys[];
 
         static const char *commandTopicTemplate;
         static const char *stateTopicTemplate;
@@ -36,20 +41,23 @@ class HAEntity
         char *uniqueId; /** Derived from id and device */
         const char *name;
         const char *component;
+        HAKVPairList features;
+        int8_t available;
 
         char *getConfigTopic(char *buffer);
         char *getConfigPayload(char *buffer, bool add_command_topic,
             bool add_state_topic);
 
-        // Build default state topic
+        /// Build default state topic
         char *getStateTopic(char *buffer);
 
 
     public:
-        HAEntity(const char *unique_id, const char *name,HADevice *device,
-            const char *component);
+        HAEntity(const char *unique_id, const char *name, const char *component,
+            HADevice *device = NULL);
 
         const char *getUniqueId();
+        inline HADevice *getDevice() { return device; }
 
         /// Some entities do not have a command topic and returns NULL
         virtual char *getCommandTopic(char *buffer) ;
@@ -68,6 +76,15 @@ class HAEntity
         // It can be redefined in the derived class
         virtual void onStateChange() {;}
 
+        /// Add predefined feature to the entity
+        void addFeature(int key, const char *value = NULL);
+        
+        /// Add not predefined feature to the entity
+        void addFeature(const char *key, const char *value);
+
+        /// It only works if previously the availability feature has been added
+        void setAvailable(bool available);
+        void sendAvailability(PubSubClient *);
 
 };