+++ src/haselect.cpp
@@ -5,7 +5,7 @@
 const char *HASelect::component PROGMEM  = "select";
 
 HASelect::HASelect(const char *unique_id, const char *name, u_int8_t max_options):
-    HAEntity(unique_id,name,NULL,component)
+    HAEntity(unique_id,name,component)
 {
     this->maxOptions = max_options;
     this->optionsAdded = 0;
@@ -14,14 +14,12 @@ HASelect::HASelect(const char *unique_id, const char *name, u_int8_t max_options
 }
 
 HASelect::HASelect(const char *unique_id,const char *name, u_int8_t max_options,
-    const char *options[]) : HAEntity(unique_id,name,NULL,component) {
+    const char *options[]) : HAEntity(unique_id,name,component) {
     this->maxOptions = max_options;
     this->state = NULL;
     this->options = (char **) options;
     this->optionsAdded = max_options;
 }
-
-
 
 HASelect::HASelect(const char *unique_id, const char *name, HADevice& device,
     u_int8_t max_options):HASelect(unique_id,name,max_options) {