+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
  - Fixed bug in unique_id when multiple devices are created
  - Improved device constructor for later setup identifier and friendly name
- - Added example for multiple devices and device
+ - Added example for multiple devices and devices configured in the setup
+   function
 
 ## [1.0.3] - 2024-04-17